use half for rand function return

+++ Shaders.metal
@@ -519,7 +519,7 @@ kernel void sampleCanvas(volatile device ColorSample *colorSample [[ buffer(0) ]
 
 
 // Generate a random float in the range [0.0f, 1.0f] using x, y, and z (based on the xor128 algorithm)
-float rand(int x, int y, int z)
+half rand(int x, int y, int z)
 {
     int seed = x + y * 57 + z * 241;
     seed= (seed<< 13) ^ seed;
